moved game actions
