+++ pages/games/[chainLabel]/[gameId].tsx
@@ -10,14 +10,12 @@ import {
   Image,
   Link,
   Spinner,
-  StackProps,
   Tab,
   TabList,
   TabPanel,
   TabPanels,
   Tabs,
   Text,
-  useDisclosure,
   VStack,
   Wrap,
 } from '@chakra-ui/react';
@@ -29,6 +27,7 @@ import { useAccount } from 'wagmi';
 import { CharacterCard } from '@/components/CharacterCard';
 import { CharactersPanel } from '@/components/CharactersPanel';
 import { ClassesPanel } from '@/components/ClassesPanel';
+import { GameActions } from '@/components/GameActions';
 import { GameTotals } from '@/components/GameTotals';
 import { ItemsPanel } from '@/components/ItemsPanel';
 import { JoinGame } from '@/components/JoinGame';
@@ -43,7 +42,6 @@ import { DropExperienceModal } from '@/components/Modals/DropExperienceModal';
 import { EditItemClaimableModal } from '@/components/Modals/EditItemClaimableModal';
 import { EquipItemModal } from '@/components/Modals/EquipItemModal';
 import { GiveItemsModal } from '@/components/Modals/GiveItemsModal';
-import { ItemsCatalogModal } from '@/components/Modals/ItemsCatalogModal';
 import { JailPlayerModal } from '@/components/Modals/JailPlayerModal';
 import { RemoveCharacterModal } from '@/components/Modals/RemoveCharacterModal';
 import { RenounceCharacterModal } from '@/components/Modals/RenounceCharacterModal';
@@ -539,46 +537,3 @@ function GamePage({
     </>
   );
 }
-
-const GameActions: React.FC<StackProps> = ({ ...props }) => {
-  const { isMaster } = useGame();
-
-  const { openActionModal } = useGameActions();
-  const itemsCatalogModal = useDisclosure();
-
-  return (
-    <VStack
-      h="100%"
-      bg="cardBG"
-      px={{ base: 4, sm: 8 }}
-      py={8}
-      align="stretch"
-      spacing={4}
-      {...props}
-    >
-      <Button onClick={itemsCatalogModal.onOpen} size="sm">
-        show items catalog
-      </Button>
-      {isMaster && (
-        <>
-          <Button
-            onClick={() => openActionModal(GameMasterActions.CREATE_ITEM)}
-            size="sm"
-          >
-            create Item
-          </Button>
-          <Button
-            onClick={() => openActionModal(GameMasterActions.CREATE_CLASS)}
-            size="sm"
-          >
-            create Class
-          </Button>
-        </>
-      )}
-      <ItemsCatalogModal
-        isOpen={itemsCatalogModal.isOpen}
-        onClose={itemsCatalogModal.onClose}
-      />
-    </VStack>
-  );
-};
